REFACTOR: lib icinde yer degisiklikleri
