+++ ulakbus/lib/date_time_helper.py
@@ -32,6 +32,10 @@ HAFTA = HAFTA_ICI_GUNLER + HAFTA_SONU_GUNLER
 
 GUN_LISTESI = ['Pazartesi', 'Salı', 'Çarşamba', 'Perşembe', 'Cuma', 'Cumartesi', 'Pazar']
 
+AYLAR = [(1, 'Ocak'), (2, 'Şubat'), (3, 'Mart'), (4, 'Nisan'),
+         (5, 'Mayıs'), (6, 'Haziran'), (7, 'Temmuz'), (8, 'Ağustos'),
+         (9, 'Eylül'), (10, 'Ekim'), (11, 'Kasım'), (12, 'Aralık')]
+
 
 def map_sinav_etkinlik_hafta_gunleri(sinavlar):
     """
